@@ -15,7 +15,12 @@
     "@types/cors": "^2.8.12",
     "@types/mongoose": "^5.11.97",
     "@types/node": "^16.11.1",
-    "eslint": "^8.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
+    "eslint": "^8.3.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.25.3",
+    "serverless-offline": "^6.9.0",
     "ts-node": "^10.3.0",
     "ts-node-dev": "^1.1.8",
     "typescript": "^4.4.4"
