@@ -1,4 +1,4 @@
-import  {User} from "./User";
-export {
-User
-};
+/* eslint-disable import/extensions */
+import User from './User';
+
+export default User;
