@@ -1,7 +1,6 @@
-import { userResolvers} from './userResolvers';
+/* eslint-disable import/extensions */
+import userResolvers from './userResolvers';
 
 const resolvers = [userResolvers];
 
-export {
-  resolvers,
-};
+export default resolvers;
