@@ -1,18 +1,16 @@
-import {User}  from "../controllers";
+/* eslint-disable import/extensions */
+import User from '../controllers';
 
 const userResolvers = {
   Query: {
   //  users: () => User.fetchUsers(),
   },
-  Mutation:{
-   signUp:(parent:any,args:any,ctx:any)=>User.signUp(parent, args,ctx),
-   //login
-   logIn:(parent:any,args:any,ctx:any)=>User.logIn(parent,args,ctx),
-
+  Mutation: {
+    signUp: (parent:any, args:any, ctx:any) => User.signUp(parent, args, ctx),
+    // login
+    logIn: (parent:any, args:any, ctx:any) => User.logIn(parent, args, ctx),
 
   }
 };
 
-export{ 
-  userResolvers,
-}
+export default userResolvers;
