@@ -1,7 +1,7 @@
-
-import { userType} from "./UserType";
-
+/* eslint-disable import/prefer-default-export */
+/* eslint-disable import/extensions */
+import userType from './UserType';
 
 export {
-userType
+  userType
 };
